lc0139: Word Break
