+++ src/main/java/com/zea7ot/lc/logs.txt
@@ -1,6 +1,6 @@
 20200708:
 3: 
-4: 0632, 1235
+4: 0139, 0632, 1235
 5: 
 
 20200707:
